cetagory section compleate

+++ src/component/Pages/MyToy/ShowMyToy/ShowMyToy.jsx
@@ -1,16 +1,36 @@
 import { Link } from "react-router-dom";
+import Swal from "sweetalert2";
 
 const ShowMyToy = ({ toy }) => {
 
     const { _id, productName, selarName, photoUrl, subCategory, price, quantity, } = toy
 
     const deletPost = _id => {
-        fetch(`https://toys-server-site.vercel.app/cameras/${_id}`, {
-            method: "DELETE"
-        })
-        .then(res => res.json())
-        .then(data => {
-            console.log(data)
+        Swal.fire({
+            title: 'Are you sure?',
+            text: "You won't be able to revert this!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Yes, delete it!'
+        }).then((result) => {
+            if (result.isConfirmed) {
+                fetch(`https://toys-server-site.vercel.app/cameras/${_id}`, {
+                    method: "DELETE"
+                })
+                    .then(res => res.json())
+                    .then(data => {
+                        console.log(data)
+                        if (data.deletedCount) {
+                            Swal.fire(
+                                'Deleted!',
+                                'Your file has been deleted.',
+                                'success'
+                            )
+                        }
+                    })
+            }
         })
     }
 
@@ -37,7 +57,7 @@ const ShowMyToy = ({ toy }) => {
                 <Link to={`/updateToy/${_id}`} className="btn btn-gray">Update</Link>
             </th>
             <th>
-                <button onClick={ () => deletPost(_id)} className="btn btn-gray">Delete</button>
+                <button onClick={() => deletPost(_id)} className="btn btn-gray">Delete</button>
             </th>
         </tr>);
 };
